breaking out init.groovy
